+++ test/TexasHoldemRoomRealKeysNoShuffleRound2.t.sol
@@ -338,7 +338,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         // the encrypted values submitted are not correct at the moment
         // submit card index and decrypted value for each card
         // card 0 to p1, card 1 to p2, card 2 to p1, card 3 to p2!
-        uint8[] memory cardIndexes = new uint8[](2);
+        uint256[] memory cardIndexes = new uint256[](2);
         cardIndexes[0] = 1;
         cardIndexes[1] = 3;
         BigNumber[] memory decryptionValues = new BigNumber[](2);
@@ -404,7 +404,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         console.log("Reveal flop stage reached");
 
         // All players should submit their decryption values for the flop cards
-        uint8[] memory cardIndexesFlop = new uint8[](3);
+        uint256[] memory cardIndexesFlop = new uint256[](3);
         BigNumber[] memory decryptionValuesFlop = new BigNumber[](3);
         // player 1 should submit their decryption values for flop cards
         vm.startPrank(player1);
@@ -471,7 +471,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         vm.stopPrank();
 
         // All players should submit their decryption values for the turn card
-        uint8[] memory cardIndexesTurn = new uint8[](1);
+        uint256[] memory cardIndexesTurn = new uint256[](1);
         BigNumber[] memory decryptionValuesTurn = new BigNumber[](1);
         // 2 players, 5th card is burned, 678 are the flop cards,
         // 9th card is burned, 10th card is the turn card
@@ -507,7 +507,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         vm.stopPrank();
 
         // All players should submit their decryption values for the river card
-        uint8[] memory cardIndexesRiver = new uint8[](1);
+        uint256[] memory cardIndexesRiver = new uint256[](1);
         BigNumber[] memory decryptionValuesRiver = new BigNumber[](1);
         // 2 players, 9th card is burned, 10th card is the turn card
         // 11th card is burned, 12th card is the river card
@@ -561,7 +561,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         );
         // also expect a winner event to be emitted
         vm.expectEmit(address(room));
-        uint8[] memory winnerPlayerIndexes = new uint8[](2);
+        uint256[] memory winnerPlayerIndexes = new uint256[](2);
         winnerPlayerIndexes[0] = 0;
         winnerPlayerIndexes[1] = 1;
         address[] memory winnerAddresses = new address[](2);
@@ -581,7 +581,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
         // the players bets should be reset and each player should have received half the pot
         //(back to starting amount)
         players = room.getPlayers();
-        for (uint8 i = 0; i < room.numPlayers(); i++) {
+        for (uint256 i = 0; i < room.numPlayers(); i++) {
             assertEq(players[i].currentStageBet, 0);
             assertEq(players[i].totalRoundBet, 0);
             assertEq(players[i].chips, 1000);
@@ -619,7 +619,7 @@ contract TexasHoldemRoomRealKeysNoShuffleRound2Test is Test {
 
         // player 2 should submit their decryption values for player 1's cards
         vm.startPrank(player2);
-        uint8[] memory cardIndexesR2 = new uint8[](2);
+        uint256[] memory cardIndexesR2 = new uint256[](2);
         cardIndexesR2[0] = 0;
         cardIndexesR2[1] = 2;
         BigNumber[] memory decryptionValuesR2 = new BigNumber[](2);
